further filtering. now we have the old cropped images
i created new data sets called final script and adjusted data_filter2.py this way it keeps all the files that has the full image and mask and makes the cropped image an optional. further more it fixed an issue where it labeled mask as crop in mass and crop as mask in calc.